+++ packages/vibrant-components/src/index.ts
@@ -60,3 +60,5 @@ export { Tooltip } from './lib/Tooltip';
 export { Table, useTable } from './lib/Table';
 export { TableHeader } from './lib/TableHeader';
 export { TableFooter, tableFooterTranslation } from './lib/TableFooter';
+export { Ratio } from './lib/Ratio';
+export { Skeleton } from './lib/Skeleton';