+++ packages/vibrant-components/src/lib/ContainedButton/index.ts
@@ -1,2 +1,3 @@
 export { ContainedButton } from './ContainedButton';
 export type { ContainedButtonProps } from './ContainedButtonProps';
+export { sizeVariation } from './constants/size';