+++ packages/vibrant-components/src/lib/Skeleton/SkeletonChip/index.ts
@@ -1,0 +1,2 @@
+export { SkeletonChip } from './SkeletonChip';
+export type { SkeletonChipProps } from './SkeletonChipProps';