+++ packages/vibrant-components/src/lib/Skeleton/SkeletonField/SkeletonFieldProps.ts
@@ -1,0 +1,26 @@
+import type { ResponsiveValue } from '@vibrant-ui/core';
+import { propVariant, withVariation } from '@vibrant-ui/core';
+
+export type SkeletonFieldProps = {
+  size: ResponsiveValue<'lg' | 'md'>;
+  width: ResponsiveValue<number | '100%'>;
+};
+
+export const withSkeletonFieldVariation = withVariation<SkeletonFieldProps>('SkeletonField')(
+  propVariant({
+    props: [
+      {
+        name: 'size',
+        responsive: true,
+      },
+    ],
+    variants: {
+      md: {
+        height: 38,
+      },
+      lg: {
+        height: 50,
+      },
+    },
+  })
+);