+++ packages/vibrant-components/src/lib/Skeleton/SkeletonField/index.ts
@@ -1,0 +1,2 @@
+export { SkeletonField } from './SkeletonField';
+export type { SkeletonFieldProps } from './SkeletonFieldProps';