+++ packages/vibrant-components/src/lib/Skeleton/SkeletonImage/index.ts
@@ -1,0 +1,2 @@
+export { SkeletonImage } from './SkeletonImage';
+export type { SkeletonImageProps } from './SkeletonImageProps';