+++ packages/vibrant-components/src/lib/Skeleton/SkeletonMotion/SkeletonMotionProps.ts
@@ -1,0 +1,8 @@
+import type { MotionProps } from 'packages/vibrant-motion/src/lib/Motion';
+import { withVariation } from '@vibrant-ui/core';
+
+export type SkeletonMotionProps = {
+  children: MotionProps['children'];
+};
+
+export const withSkeletonMotionVariation = withVariation<SkeletonMotionProps>('SkeletonMotion')();