+++ packages/vibrant-core/src/lib/Text/TextProps.ts
@@ -1,3 +1,4 @@
+import type { Ref } from 'react';
 import type { ReactElementChild, ReactTextChild } from '../../types';
 import { createInterpolation } from '../createInterpolation';
 import { createShouldForwardProp } from '../createShouldForwardProp';
@@ -55,14 +56,13 @@ export type TextProps = SystemProps & {
   as?: TextElements;
   children?: TextChildren;
   onCopy?: () => void;
+  ref?: Ref<any>;
 };
 
 export const shouldForwardProp = createShouldForwardProp(systemPropNames);
 
 export const interpolation = injectContext(
   createInterpolation(systemProps, {
-    display: 'flex',
-    flexDirection: 'column',
     textAlign: 'left',
     color: 'onColor',
   })