+++ packages/vibrant-components/CHANGELOG.md
@@ -2,6 +2,43 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.37.1](https://github.com/pedaling/opensource/compare/vibrant-components-0.37.0...vibrant-components-0.37.1) (2023-01-06)
+
+
+### Features
+
+* add custom ref into Tooltip ([#519](https://github.com/pedaling/opensource/issues/519)) ([388ddb5](https://github.com/pedaling/opensource/commit/388ddb5e4bf1e615675cf3924e49fca26a6e2da0))
+
+
+### Bug Fixes
+
+* remove default display in Text ([#518](https://github.com/pedaling/opensource/issues/518)) ([6d60f24](https://github.com/pedaling/opensource/commit/6d60f240a6240e47487a3d17c3a052265c9f8b5a))
+
+## [0.37.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.36.2...vibrant-components-0.37.0) (2023-01-05)
+
+
+### Features
+
+* add Ratio component ([#513](https://github.com/pedaling/opensource/issues/513)) ([95111be](https://github.com/pedaling/opensource/commit/95111be5b351cd09f07e9ae694527baf91476e2b))
+* add Skeleton component ([#511](https://github.com/pedaling/opensource/issues/511)) ([b3eb7c3](https://github.com/pedaling/opensource/commit/b3eb7c37d61c29e61a56144e79bf5848121c377b))
+
+
+### Bug Fixes
+
+* cannot read properties of null issue ([#516](https://github.com/pedaling/opensource/issues/516)) ([6016722](https://github.com/pedaling/opensource/commit/6016722623dc63631492e587297a06f30e60ff85))
+
+### [0.36.2](https://github.com/pedaling/opensource/compare/vibrant-components-0.36.1...vibrant-components-0.36.2) (2023-01-05)
+
+
+### Features
+
+* can use boolean or null type in Table children ([#515](https://github.com/pedaling/opensource/issues/515)) ([f856f5e](https://github.com/pedaling/opensource/commit/f856f5e39391a4d9e19793a583b491907e54cd9f))
+
+
+### Bug Fixes
+
+* Tooltip 위치 조정을 수정한다 ([#514](https://github.com/pedaling/opensource/issues/514)) ([62085f4](https://github.com/pedaling/opensource/commit/62085f4003ea6f63470c0681d29fede17ba24f3d))
+
 ### [0.36.1](https://github.com/pedaling/opensource/compare/vibrant-components-0.36.0...vibrant-components-0.36.1) (2023-01-03)