+++ packages/vibrant-core/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.37.1](https://github.com/pedaling/opensource/compare/vibrant-core-0.37.0...vibrant-core-0.37.1) (2023-01-06)
+
+
+### Features
+
+* add custom ref into Tooltip ([#519](https://github.com/pedaling/opensource/issues/519)) ([388ddb5](https://github.com/pedaling/opensource/commit/388ddb5e4bf1e615675cf3924e49fca26a6e2da0))
+
+
+### Bug Fixes
+
+* remove default display in Text ([#518](https://github.com/pedaling/opensource/issues/518)) ([6d60f24](https://github.com/pedaling/opensource/commit/6d60f240a6240e47487a3d17c3a052265c9f8b5a))
+
+## [0.37.0](https://github.com/pedaling/opensource/compare/vibrant-core-0.36.2...vibrant-core-0.37.0) (2023-01-05)
+
+
+### Features
+
+* add Ratio component ([#513](https://github.com/pedaling/opensource/issues/513)) ([95111be](https://github.com/pedaling/opensource/commit/95111be5b351cd09f07e9ae694527baf91476e2b))
+* add Skeleton component ([#511](https://github.com/pedaling/opensource/issues/511)) ([b3eb7c3](https://github.com/pedaling/opensource/commit/b3eb7c37d61c29e61a56144e79bf5848121c377b))
+
+### [0.36.2](https://github.com/pedaling/opensource/compare/vibrant-core-0.36.1...vibrant-core-0.36.2) (2023-01-05)
+
 ### [0.36.1](https://github.com/pedaling/opensource/compare/vibrant-core-0.36.0...vibrant-core-0.36.1) (2023-01-03)